# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "DATN E-Commerce API",
        "description": "Tài liệu API hệ thống DATN E-Commerce Backend",
        "contact": {
            "email": "admin@example.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://127.0.0.1:8000",
            "description": "DATN API Server"
        }
    ],
    "paths": {
...
        "/api/admin/orders": {
            "get": {
                "tags": [
                    "Đơn hàng (Order)"
                ],
                "summary": "[Admin] Danh sách tất cả đơn hàng",
                "operationId": "4cb77c0a02289742e85d592a2584da08",
                "responses": {
                    "200": {
                        "description": "Thành công"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/admin/orders/{id}/status": {
            "post": {
                "tags": [
                    "Đơn hàng (Order)"
                ],
                "summary": "[Admin] Cập nhật trạng thái đơn hàng",
                "operationId": "9d8065b0bdd28c3892c8af09e303e688",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "status"
                                ],
                                "properties": {
                                    "status": {
                                        "type": "string",
                                        "enum": [
                                            "new",
                                            "pending",
                                            "shipping",
                                            "delivered",
                                            "cancelled"
                                        ]
                                    },
                                    "payment_status": {
                                        "type": "string",
                                        "enum": [
                                            "pending",
                                            "paid",
                                            "refunded"
                                        ]
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Cập nhật thành công"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/admin/orders/{id}": {
            "delete": {
                "tags": [
                    "Đơn hàng (Order)"
                ],
                "summary": "[Admin] Xóa đơn hàng",
                "operationId": "8febd1bde00f9b437095f794c6c9e0eb",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Xóa thành công"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/user/orders": {
            "get": {
                "tags": [
                    "Đơn hàng (Order)"
                ],
                "summary": "Danh sách đơn mua của tôi",
                "operationId": "836184aa55de189c467e97c41a21f4eb",
                "responses": {
                    "200": {
                        "description": "Thành công"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/user/orders/{id}/cancel": {
            "post": {
                "tags": [
                    "Đơn hàng (Order)"
                ],
                "summary": "Hủy đơn hàng của tôi",
                "operationId": "7e7dfe3a529604d55f7306987d020453",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "reason": {
                                        "type": "string",
                                        "example": "Đổi ý không muốn mua nữa"
                                    },
                                    "restore_cart": {
                                        "type": "boolean",
                                        "example": false
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Hủy thành công"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/user/orders/{id}/confirm-payment": {
            "post": {
                "tags": [
                    "Đơn hàng (Order)"
                ],
                "summary": "Xác nhận thanh toán đơn hàng",
                "operationId": "f499e2950296c3930f5eb724c35a0348",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Xác nhận thành công"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/user/orders/{id}": {
            "get": {
                "tags": [
                    "Đơn hàng (Order)"
                ],
                "summary": "Chi tiết một đơn hàng của tôi",
                "operationId": "b2366769f8c470169519863083551275",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Thành công"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Đơn hàng (Order)"
                ],
                "summary": "Hủy và khôi phục giỏ hàng cho đơn chưa thanh toán",
                "operationId": "a2b2fd0176ee520081d897041fc14fed",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Xóa thành công"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "description": "Nhập token của bạn dạng: Bearer {token}",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}